refactor: moves data planes into app directory
