+++ src/app/data-planes/components/DataPlaneDetails.vue
