+++ src/app/data-planes/components/DataPlaneEntitySummary.vue
