+++ src/app/data-planes/views/DataPlaneDetailView.vue
